+++ ElementX/Sources/Services/BugReport/BugReportService.swift
@@ -342,14 +342,3 @@ extension BugReportService: URLSessionTaskDelegate {
             .store(in: &cancellables)
     }
 }
-
-extension URLSession {
-    /// The same as `data(for:delegate:)` but with an additional immediate retry if the first attempt fails.
-    func dataWithRetry(for request: URLRequest, delegate: URLSessionTaskDelegate? = nil) async throws -> (Data, URLResponse) {
-        if let firstTryResult = try? await data(for: request, delegate: delegate) {
-            return firstTryResult
-        }
-        
-        return try await data(for: request, delegate: delegate)
-    }
-}
